plutus-pab: split Main
Split the (pretty big) Main module up:

- Command : data type representing all available commands
- CommandParser: command line parser for all commands
- Cli: command interpreter for all commands
