+++ plutus-pab/plutus-pab.cabal
@@ -203,6 +203,9 @@ executable plutus-pab
     hs-source-dirs: app
     other-modules:
         PSGenerator
+        Cli
+        Command
+        CommandParser
     default-language: Haskell2010
     ghc-options: -threaded -rtsopts -with-rtsopts=-N -Wall -Wcompat
                  -Wincomplete-uni-patterns -Wincomplete-record-updates
